Revert "Merge branch 'master' of https://github.com/Arquisoft/dede_es5c"
This reverts commit fae098774803224532daee0dc2eced2eb247b530, reversing
changes made to f1531f7995e1d7971b03446823f1ca49cd91c22c.
