+++ restapi/api.ts
@@ -1,15 +1,9 @@
 import express, { Request, Response, Router } from 'express';
 import {check} from 'express-validator';
-import {findProductByName, findProducts, addProducto} from "./controllers/ProductController";
-import {
-    addPedido,
-    calculatePrice,
-    findPedidoByClientEmail,
-    findPedidoByWebid,
-    findPedidos
-} from "./controllers/PedidoController";
-import {createUser, findUsers, findUsersByEmail, deleteUser} from "./controllers/UserController";
-import {addOrderProduct, findAllOrderProducts, findAllProductsForOrder} from "./controllers/PedidoProductoController";
+import {findProductByName, findProducts, addProducto} from "./models/controllers/ProductController";
+import {addPedido, findPedidos} from "./models/controllers/PedidoController";
+import {createUser, findUsers, findUsersByEmail, deleteUser} from "./models/controllers/UserController";
+import {addOrderProduct, findAllOrderProducts} from "./models/controllers/PedidoProductoController";
 
 const bodyParser = require('body-parser')
 
@@ -61,10 +55,6 @@ api.get("/products/:name", findProductByName)
 //-------pedidos
 api.post("/pedido/add", addPedido)
 api.get("/pedido/list", findPedidos)
-api.get("/pedido/:webid", findPedidoByWebid)
-api.get("/pedido/:email", findPedidoByClientEmail)
-api.get("/pedido/price", calculatePrice)
-
 
 //-------usuarios
 api.get("/users/list", findUsers)
@@ -75,8 +65,6 @@ api.post("/users/delete/:email", deleteUser)
 //-------pedido-producto
 api.post("/pedidoProducto/add",addOrderProduct)
 api.get("/pedidoProducto/list",findAllOrderProducts)
-api.get("/pedidoProducto/list/:id_pedido",findAllProductsForOrder)
-
