+++ webapp/server.ts
@@ -2,10 +2,8 @@ import express,{Application} from 'express';
 //for using an import here we need to configure the tsconfig.json
 //setting the option module to commonjs
 
-//process.env.PORT
-
 var app: Application = express()
-const port: number = process.env.PORT;
+const port: number = 3000;
 
 app.use(express.static('build'))
